ml-service/requirements.txt
Flask
tensorflow
numpy
urllib3
Pillow